restructured commandlets
