+++ packages/salesforcedx-vscode-core/package.json
@@ -29,6 +29,7 @@
     "@salesforce/kit": "1.2.2",
     "@salesforce/salesforcedx-utils-vscode": "49.0.0",
     "@salesforce/source-deploy-retrieve": "1.0.15",
+    "@salesforce/salesforcedx-apex": "github:forcedotcom/salesforcedx-apex#ananya/libraryExecute",
     "adm-zip": "0.4.13",
     "applicationinsights": "1.0.7",
     "glob": "^7.1.2",
