+++ packages/salesforcedx-vscode-core/src/commands/util/index.ts
@@ -39,3 +39,4 @@ export {
   DevUsernameChecker,
   EmptyPreChecker
 } from './preconditionCheckers';
+export { DeployRetrieveLibraryExecutor } from './deployRetrieveCommandlet';
