+++ packages/salesforcedx-vscode-core/src/commands/util/libraryCommandlet.ts
@@ -25,95 +25,14 @@ import { CommandletExecutor } from './sfdxCommandlet';
 
 export abstract class LibraryCommandletExecutor<T>
   implements CommandletExecutor<T> {
-  public static errorCollection = languages.createDiagnosticCollection(
-    'deploy-errors'
-  );
   protected showChannelOutput = true;
-  protected sourceClient: SourceClient | undefined;
   protected executionName: string = '';
   protected startTime: [number, number] | undefined;
   protected telemetryName: string | undefined;
 
+  public build(execName: string, telemetryLogName: string) {}
+
   public execute(response: ContinueResponse<T>): void {}
-
-  public async build(
-    execName: string,
-    telemetryLogName: string
-  ): Promise<void> {
-    this.executionName = execName;
-    this.telemetryName = telemetryLogName;
-    // initialize connection
-    const usernameOrAlias = await OrgAuthInfo.getDefaultUsernameOrAlias(true);
-    if (!usernameOrAlias) {
-      throw new Error(nls.localize('error_no_default_username'));
-    }
-    const conn = await OrgAuthInfo.getConnection(usernameOrAlias);
-    // @ts-ignore private logger mismatch
-    this.sourceClient = new SourceClient(conn);
-  }
-
-  public retrieveWrapper(fn: (...args: any[]) => Promise<ApiResult>) {
-    const commandName = this.executionName;
-
-    return async function(...args: any[]): Promise<ApiResult> {
-      channelService.showCommandWithTimestamp(`Starting ${commandName}`);
-
-      const result = await window.withProgress(
-        {
-          title: commandName,
-          location: ProgressLocation.Notification
-        },
-        async () => {
-          // @ts-ignore
-          return (await fn.call(this, ...args)) as ApiResult;
-        }
-      );
-
-      channelService.appendLine(outputRetrieveTable(result));
-      channelService.showCommandWithTimestamp(`Finished ${commandName}`);
-      await notificationService.showSuccessfulExecution(commandName);
-      return result;
-    };
-  }
-
-  public deployWrapper(fn: (...args: any[]) => Promise<DeployResult>) {
-    const commandName = this.executionName;
-
-    return async function(...args: any[]): Promise<DeployResult> {
-      channelService.showCommandWithTimestamp(`Starting ${commandName}`);
-
-      const result = await window.withProgress(
-        {
-          title: commandName,
-          location: ProgressLocation.Notification
-        },
-        async () => {
-          // @ts-ignore
-          return (await fn.call(this, ...args)) as DeployResult;
-        }
-      );
-
-      const parser = new LibraryDeployResultParser(result);
-      const outputResult = await parser.outputResult();
-      channelService.appendLine(outputResult);
-      channelService.showCommandWithTimestamp(`Finished ${commandName}`);
-
-      if (
-        result.State === DeployStatusEnum.Completed ||
-        result.State === DeployStatusEnum.Queued
-      ) {
-        LibraryCommandletExecutor.errorCollection.clear();
-        await notificationService.showSuccessfulExecution(commandName);
-      } else {
-        handleLibraryDiagnostics(
-          result,
-          LibraryCommandletExecutor.errorCollection
-        );
-        notificationService.showFailedExecution(commandName);
-      }
-      return result;
-    };
-  }
 
   public logMetric(properties?: Properties, measurements?: Measurements) {
     telemetryService.sendCommandEvent(
